initial commit
Initial commit implements a one-off ztunnel CNI shim.

This shim implements the minimal ZDS protocol with ztunnel, over the
canonical unix socket, necessary to have ztunnel create sockets in
a child namespace.

Because this is simple PoC code the CNI shim supports only adding
ztunnel sockets into a single network namespace, provided to the CNI
shim via the sole cli argument. This argument is simply a path to
a pinned networknamespace on the file system.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>